Took care of flag titles, encoding in the wp_editor, inserting pictures into the media field, and a datepicker. Also cleaned up the look a bit, seperating the fields into columns. Next is general QA. Also reorganized the codebase a little bit.
