+++ metabox/MetaBox.js
@@ -1,0 +1,3 @@
+jQuery(document).ready(function($){
+	$( ".datepicker" ).datepicker({ altFormat: "yyyy,mm,dd" });
+});
