+++ timeline-metabox.php
@@ -3,11 +3,16 @@
 add_action( 'add_meta_boxes', 'timeline_metabox_init' );
 
 function timeline_metabox_init() {
-		wp_register_style(  'metabox', VERITETIMELINE_URL . '/wpalchemy/MetaBox.css' );
-		wp_enqueue_style( 'metabox' );
+		wp_register_style( 'timeline-metaboxcss', VERITETIMELINE_URL . '/wpalchemy/MetaBox.css' );
+		wp_register_style( 'timeline-datepickercss', VERITETIMELINE_URL . '/wpalchemy/jquery-ui-datepicker.css' );
+		wp_register_script( 'timeline-metaboxjs', VERITETIMELINE_URL . '/wpalchemy/MetaBox.js', array( 'jquery', 'jquery-ui-datepicker' ), null, true);
+		wp_enqueue_style( 'timeline-metaboxcss' );
+		wp_enqueue_style( 'timeline-datepickercss' );
+		wp_enqueue_script( 'timeline-metaboxjs' );
+		wp_enqueue_script( 'jquery-ui-datepicker' );
 }
 
-// We do class_exists in case another plugin has wpalchemy
+// We do class_exists in case another plugin uses wpalchemy
 if ( ! class_exists( 'WPAlchemy_MetaBox' ) )
 include_once VERITETIMELINE_DIR . '/wpalchemy/MetaBox.php';
 
@@ -23,7 +28,7 @@ $wpalchemy_media_access = new WPAlchemy_MediaAccess();
 
 $timeline_mb = new WPAlchemy_Metabox( array(
 	'id' => '_timeline_mb',
-	'title' => 'Timeline Extras',
+	'title' => __( 'Timeline', 'verite-timeline' ),
 	'types' => array( 'timeline' ),
 	'template' => VERITETIMELINE_DIR . '/wpalchemy/metabox-timeline.php'
 ));
